Keep iterating on the server connections
