+++ lttnetworking/src/lib.rs
@@ -7,6 +7,7 @@ pub mod messages;
 pub mod server;
 
 mod supported_games;
+mod example_supported_games;
 pub use supported_games::SupportedGames;
 
 mod user;
