+++ lttnetworking/src/server/mod.rs
@@ -1,1 +1,2 @@
 pub mod server_connection;
+pub mod server_sub_connection;
